bestfitting/requirements.txt
tqdm==4.15.0
requests==2.21.0
numpy==1.14.6
pandas==0.23.1
Pillow==5.3.0
scipy==0.19.1
ImageHash==4.0
iterative-stratification==0.1.6
torch==0.4.1
torchvision==0.2.1
opencv-python==3.4.3.18
mlcrate==0.1.0
multiprocess==0.70.5
scikit-image==0.13.0
scikit-learn==0.19.0
